+++ lib/ui/screens/overview_screen.dart
@@ -90,6 +90,7 @@ class OverviewScreen extends StatelessWidget {
             color: ZOColors.onPrimaryLight,
           ),
         ),
+        backgroundColor: ZOColors.successLight,
       ),
     );
   }
@@ -101,12 +102,13 @@ class OverviewScreen extends StatelessWidget {
         infoValue: const DonationCountdownTimer(),
         button: ZOButton(
           text: context.l10n!.callACourier,
-          icon: Icons.directions_car_filled_outlined,
           fullWidth: false,
+          type: ZOButtonType.success,
           onPressed: () async {
             await DeliveryUtils.confirmDelivery(context);
           },
         ),
+        backgroundColor: ZOColors.successLight,
       ),
     );
   }