+++ lib/ui/widgets/button.dart
@@ -3,47 +3,59 @@ import 'package:zachranobed/common/constants.dart';
 
 class ZOButton extends StatelessWidget {
   final String text;
-  final IconData icon;
   final VoidCallback onPressed;
-  final bool isSecondary;
+  final ZOButtonType type;
+  final IconData? icon;
   final double height;
   final bool fullWidth;
 
   const ZOButton({
     super.key,
     required this.text,
-    required this.icon,
     required this.onPressed,
-    this.isSecondary = false,
+    this.type = ZOButtonType.primary,
+    this.icon,
     this.height = 56.0,
     this.fullWidth = true,
   });
 
   @override
   Widget build(BuildContext context) {
-    final backgroundColor = isSecondary ? ZOColors.secondary : ZOColors.primary;
-    final foregroundColor =
-        isSecondary ? ZOColors.onSecondary : ZOColors.onPrimary;
-
-    return ElevatedButton.icon(
-      style: ElevatedButton.styleFrom(
-        backgroundColor: backgroundColor,
-        minimumSize: fullWidth ? Size.fromHeight(height) : null,
-        shape: const StadiumBorder(),
-      ),
-      onPressed: onPressed,
-      icon: Icon(
-        icon,
-        size: 18.0,
-        color: foregroundColor,
-      ),
-      label: Text(
-        text,
-        style: TextStyle(
-          color: foregroundColor,
-          fontSize: FontSize.xs,
-        ),
-      ),
+    final style = ElevatedButton.styleFrom(
+      foregroundColor: type.foregroundColor,
+      backgroundColor: type.backgroundColor,
+      minimumSize: fullWidth ? Size.fromHeight(height) : null,
+      shape: const StadiumBorder(),
+      textStyle: const TextStyle(fontSize: FontSize.xs),
     );
+
+    if (icon != null) {
+      return ElevatedButton.icon(
+        style: style,
+        onPressed: onPressed,
+        icon: Icon(
+          icon,
+          size: 18.0,
+        ),
+        label: Text(text),
+      );
+    } else {
+      return ElevatedButton(
+        style: style,
+        onPressed: onPressed,
+        child: Text(text),
+      );
+    }
   }
+}
+
+enum ZOButtonType {
+  primary(ZOColors.primary, ZOColors.onPrimary),
+  secondary(ZOColors.secondary, ZOColors.onSecondary),
+  success(ZOColors.success, ZOColors.onSuccess);
+
+  const ZOButtonType(this.backgroundColor, this.foregroundColor);
+
+  final Color backgroundColor;
+  final Color foregroundColor;
 }